+++ src/index.ts
@@ -1,3 +1,4 @@
+import autocomplete from "./autocomplete";
 import search from "./search";
 import random from "./random";
 import getAllCombos from "./get-all-combos";
@@ -6,6 +7,7 @@ import makeFakeCombo from "./make-fake-combo";
 import { resetCache } from "./spellbook-api";
 
 export = {
+  autocomplete,
   search,
   random,
   getAllCombos,