+++ test/integration/commander-spellbook.test.ts
@@ -369,4 +369,60 @@ describe("Commander Spellbook", () => {
       expect(combos[0].results).toBeTruthy();
     });
   });
+
+  describe("autocomplete", () => {
+    it("returns all card names", async () => {
+      const names = await spellbook.autocomplete("cards", "");
+
+      expect(names.length).toBeGreaterThan(2500);
+    });
+
+    it("returns card names that match partially", async () => {
+      const names = await spellbook.autocomplete("cards", "mind");
+
+      expect(names.length).toBeLessThan(50);
+      names.forEach((name) => {
+        expect(name.value).toContain("mind");
+      });
+    });
+
+    it("returns card names that match exactly (regardless of capitalization and puntuation)", async () => {
+      const names = await spellbook.autocomplete(
+        "cards",
+        "sydri galvanic genius"
+      );
+
+      expect(names.length).toBe(1);
+      expect(names[0].value).toBe("sydri galvanic genius");
+      expect(names[0].label).toBe("Sydri, Galvanic Genius");
+    });
+
+    it("returns all results", async () => {
+      const results = await spellbook.autocomplete("results", "");
+
+      expect(results.length).toBeGreaterThan(700);
+    });
+
+    it("returns results that match partially", async () => {
+      const results = await spellbook.autocomplete("results", "infinite");
+
+      expect(results.length).toBeLessThan(700);
+      results.forEach((result) => {
+        expect(result.value.toLowerCase()).toContain("infinite");
+      });
+    });
+
+    it("returns all colors", async () => {
+      const colors = await spellbook.autocomplete("colors", "");
+
+      // 32 + 5 entries for sans-color
+      expect(colors.length).toBe(37);
+    });
+
+    it("returns colors that match partially", async () => {
+      const colors = await spellbook.autocomplete("colors", "sans");
+
+      expect(colors.length).toBe(5);
+    });
+  });
 });